metrics: move the surface test next to the other metric tests
It went into a file of its own when it was written; metrics_test.go is where
the rest of them live.
